10th build of the 24h marathon

+++ src/main/java/com/portalthree/jed/handlers/TextRenderer.java
@@ -8,7 +8,7 @@ import org.lwjgl.opengl.GL11;
 
 public class TextRenderer extends Gui {
     public TextRenderer(Minecraft mc, String text, int x, int y, double scale) {
-        double scaleReset = (double) Math.pow(scale, -1);
+        double scaleReset = Math.pow(scale, -1);
 
         GL11.glScaled(scale, scale, scale);
         y -= mc.fontRendererObj.FONT_HEIGHT;
